[threads/stencil] deserialize execution of communication pattern
Currently, data exchange with the north is initiated only after
the exchange with the south is completed. This causes an
unnecessary cascade of dependence.

This commit initiates all the operations in an iteration before
waiting for them to complete. The right use of communicator,
rank, and tag ensures correct matching between the issued
operations.